Add resize option, cleanup of tasks
